backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-parler==2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
geopandas==1.0.1
googlemaps==4.10.0
idna==3.7
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pip-review==1.3.0
psycopg==3.2.1
psycopg-binary==3.2.1
PyJWT==2.8.0
pyogrio==0.9.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
shapely==2.0.5
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.2
